% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoa_gbif.R
\name{hoa_gbif}
\alias{hoa_gbif}
\title{GBIF}
\usage{
hoa_gbif(x, limit = 500, ...)
}
\arguments{
\item{x}{(character) a scientific name}

\item{limit}{(numeric) number records to get. default: 500}

\item{...}{curl options}
}
\description{
GBIF
}
\examples{
\dontrun{
res <- hoa_gbif(x = "Remora remora")
res$occ_remarks
res$ev_remarks
library(dplyr)
full_join(res$occ_remarks, res$ev_remarks, by = "key")
}
}

