% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoa_search.R
\name{hoa_search}
\alias{hoa_search}
\title{Search}
\usage{
hoa_search(x, from = "gbif", limit = 500, ...)
}
\arguments{
\item{x}{(character) a scientific name}

\item{from}{(character) One of 'gbif', 'vertnet' or others (to come later)}

\item{limit}{(numeric) number records to get. default: 500}

\item{...}{curl options}
}
\description{
Search
}
\examples{
\dontrun{
hoa_search(x = "Remora remora", from = "gbif")
}
}

